adding utility bars and auto routing on non signed in users
