+++ src/components/element-components/ChatNav.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Box, Typography } from "@mui/material";
+
+const ChatNav = () => {
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        border: "1px solid white",
+        justifyContent: "space-between",
+        alignItems: "center",
+        width: "90vw",
+      }}
+    >
+      <Typography>ghello</Typography>
+    </Box>
+  );
+};
+
+export default ChatNav;
