+++ src/components/element-components/InputArea.jsx
@@ -3,7 +3,6 @@ import React, { useContext, useState } from "react";
 import { db } from "../../../server/firebaseConfig";
 import { collection, addDoc, serverTimestamp } from "firebase/firestore";
 import { UserContext } from "../../App";
-import { askTimmy } from "../utils/AskTimmy";
 
 const InputArea = () => {
   const [message, setMessage] = useState("");
@@ -33,14 +32,6 @@ const InputArea = () => {
     setMessage("");
   };
 
-  const handleAskTimmy = (e) => {
-    e.preventDefault();
-
-    askTimmy("hi man");
-
-    setMessage("");
-  };
-
   return (
     <Box sx={{ width: "100%", height: "20%", border: "1px solid red" }}>
       <form onSubmit={handleSendMessage}>
@@ -52,11 +43,8 @@ const InputArea = () => {
           value={message}
           sx={{ input: { color: "white" }, width: "70%" }}
         />
-        <Button sx={{ width: "15%" }} type="submit">
+        <Button sx={{ width: "25%" }} type="submit">
           Send Message
-        </Button>
-        <Button sx={{ width: "15%" }} onClick={handleAskTimmy}>
-          Ask Timmy
         </Button>
       </form>
     </Box>
