+++ src/components/page-components/About.jsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react";
 const About = () => {
   return (
     <div>
-      <h2>Ask Timmy</h2>
+      <h2>Chat</h2>
       <p>Hello</p>
     </div>
   );
